Blank values in forms should be treated as None
Right now if you have a form field that is optional and you leave it blank, it throws an error because the value it is getting is "" which is not a valid JSON string. I'm not sure if the value is *supposed* to always be None when you leave a field blank and it isn't in my app for some reason, but even so I think there's definitely a difference between "entered an incorrect value for JSON" and "didn't enter anything and the value is an empty string". In the latter case, it should be treated identically to None.